[Cleanup] Merging PMonitor with Monitor
